backend/requirements.txt
# ─── Heaven Coint Backend Dependencies ─────────────────────
# Python 3.12+

# Web framework
fastapi==0.115.12
uvicorn[standard]==0.34.2
python-multipart==0.0.20

# Database
sqlalchemy[asyncio]==2.0.40
asyncpg==0.30.0
psycopg2-binary==2.9.10
alembic==1.15.2

# Redis
redis[hiredis]==5.3.0

# Auth
python-jose[cryptography]==3.4.0
passlib[bcrypt]==1.7.4
bcrypt==4.3.0

# Validation / Settings
pydantic==2.11.3
pydantic-settings==2.9.1
email-validator==2.2.0

# HTTP client (for external APIs)
httpx==0.28.1
